+++ docs/CHANGELOG.md
@@ -1,3 +1,12 @@
+
+## [9.5.24] - 2026-01-29
+- Automated release build.
+
+## [9.5.23] - 2026-01-29
+- Automated release build.
+
+## [9.5.22] - 2026-01-29
+- Automated release build.
 # Changelog
 
 ## [9.5.21] - 2026-01-28